+++ src/data/+mag/+meta/Science.m
@@ -1,10 +1,6 @@
 classdef Science < mag.meta.Data
 % SCIENCE Description of MAG science data.
 
-    properties (Constant, Hidden)
-        MetaDataFilePattern (1, 1) string = "MAGScience-(?<mode>\w+)-\((?<primaryFrequency>\d+),(?<secondaryFrequency>\d+)\)-(?<packetFrequency>\d+)s-(?<date>\d+)-(?<time>\w+).(?<extension>\w+)"
-    end
-
     properties
         % MODEL Sensor model type and number.
         Model string {mustBeScalarOrEmpty, mag.validator.mustMatchRegex(Model, "^[LEF]M\d$")}